Move hooks to new file
